feat #37: do minor quality improvements in Predictor/Preprocessor

+++ app/src/main/java/org/htwk/pacing/backend/predictor/Predictor.kt
@@ -27,7 +27,7 @@ class Predictor {
 
     data class FixedParameters(
         //TODO: add more fixed vital parameters
-        val anaerobicThreshold: Float
+        val anaerobicThreshold: Double
     )
 
     fun run(
